+++ src/System.Private.ServiceModel/tests/Scenarios/Security/TransportSecurity/Tcp/IdentityTests.cs
@@ -1,4 +1,4 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
@@ -18,8 +18,8 @@ public partial class IdentityTests : ConditionalWcfTest
     [ConditionalFact(nameof(Root_Certificate_Installed), nameof(Client_Certificate_Installed))]
 #endif
     [OuterLoop]
-    // The product code will check the Dns identity from the server and throw if it does not match what is specified in DnsEndpointIdentity
-    public static void VerifyServiceIdentityMatchDnsEndpointIdentity()
+    // Verify product throws MessageSecurityException when the Dns identity from the server does not match the expectation
+    public static void ServiceIdentityNotMatch_Throw_MessageSecurityException()
     {
 #if FULLXUNIT_NOTSUPPORTED
         bool root_Certificate_Installed = Root_Certificate_Installed();
@@ -39,16 +39,19 @@ public partial class IdentityTests : ConditionalWcfTest
         binding.Security.Mode = SecurityMode.Transport;
         binding.Security.Transport.ClientCredentialType = TcpClientCredentialType.None;
 
-        EndpointAddress endpointAddress = new EndpointAddress(new Uri(Endpoints.Tcp_VerifyDNS_Address), new DnsEndpointIdentity(Endpoints.Tcp_VerifyDNS_HostName));
+        EndpointAddress endpointAddress = new EndpointAddress(new Uri(Endpoints.Tcp_VerifyDNS_Address), new DnsEndpointIdentity("wrongone"));
         ChannelFactory<IWcfService> factory = new ChannelFactory<IWcfService>(binding, endpointAddress);
         IWcfService serviceProxy = factory.CreateChannel();
 
         try
         {
-            var result = serviceProxy.Echo(testString);
-            Assert.Equal(testString, result);
+            var exception = Assert.Throws<MessageSecurityException>(() =>
+            {
+                var result = serviceProxy.Echo(testString);
+                Assert.Equal(testString, result);
+            });
 
-            factory.Close();
+            Assert.True(exception.Message.Contains(Endpoints.Tcp_VerifyDNS_HostName), string.Format("Expected exception message contains: '{0}', actual: '{1}')", Endpoints.Tcp_VerifyDNS_HostName, exception.Message));
         }
         finally
         {
